+++ spotifyutils/__init__.py
@@ -1,20 +1,7 @@
 import argparse
 import os
-import json
-import http.server
-import socketserver
-import webbrowser
-import urllib.request
-import urllib.parse
 import configparser
-import tempfile
-import base64
-import ssl
-from spotifyutils.cert import generate_selfsigned_cert
-
-configfile = ''
-REDIRECT_URI = ''
-AUTHCODE = ''
+from spotifyutils.auth import userAuth, server, getTokens
 
 def cli(input_args=None):
     """Main playlist program entrypoint
@@ -68,8 +55,6 @@ def cli(input_args=None):
 def main(**kwargs):
     """ Direct user to authorize spotify URL """
 
-    global REDIRECT_URI
-    global configfile
     configfile = kwargs['configfile']
     
     config = configparser.ConfigParser()
@@ -84,106 +69,14 @@ def main(**kwargs):
 
     parser = configparser.ConfigParser()
     parser.read(configfile)
-    REDIRECT_URI = parser.get('spotifyutils', 'redirect_uri')
-    baseUrl = "https://accounts.spotify.com/authorize?"
-    response_type = 'code'
-    scope = 'user-read-email'
-    show_dialog = 'true'
 
-    PARAMS = {
-        'client_id': parser.get('spotifyutils', 'client_id'),
-        'response_type': response_type,
-        'redirect_uri': REDIRECT_URI,
-        'scope': scope,
-        'show_dialog': show_dialog
-    }
+    client_id = parser.get('spotifyutils', 'client_id')
+    client_secret = parser.get('spotifyutils', 'client_secret')
+    redirect_uri = parser.get('spotifyutils', 'redirect_uri')
 
-    authorizationURL = (baseUrl + (urllib.parse.urlencode(PARAMS)))
-    webbrowser.open_new(authorizationURL)
+    print("Redirecting to Spotify Authorization URI, and spinning up web server")
+    userAuth(redirect_uri, client_id)
+    server()
 
-
-class WebServer(http.server.BaseHTTPRequestHandler):
-
-    def do_GET(self):
-        """ Handle auth code sent to redirect URI """
-        global AUTHCODE
-        
-        # Read URI from configfile
-        uri = urllib.parse.urlparse(REDIRECT_URI).path
-
-        # Parse path out of GET request
-        endpoint = urllib.parse.urlparse(self.path).path
-
-        if uri == endpoint:
-            params = urllib.parse.parse_qs(urllib.parse.urlparse(self.path).query)
-            AUTHCODE = params['code'][0]
-
-            self.path = 'spotifyutils/index.html'
-            indexFile = open(self.path).read()
-            self.send_response(200, 'OK')
-            self.end_headers()
-            self.wfile.write(bytes(indexFile, 'utf-8'))
-
-        else:
-            self.send_response(404, 'NOT FOUND')
-
-def secureServer(http_server, host):
-    """ generate items required to make web server secure """
-
-    ssl_context = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
-    
-    # create a temporary, throwaway certificate
-    cert, private = generate_selfsigned_cert(host)
-
-    # these files are deleted after the context exists
-    with tempfile.NamedTemporaryFile() as cert_file:
-        with tempfile.NamedTemporaryFile() as key_file:
-
-            cert_file.write(cert)
-            key_file.write(private)
-            cert_file.flush()
-            key_file.flush()
-
-            ssl_context.load_cert_chain(
-                cert_file.name, key_file.name
-            )
-
-    http_server.socket = ssl_context.wrap_socket(http_server.socket, server_side=True) 
-    return http_server
-        
-def server():
-    """ create a web server to handle the get request """
-    
-    host, port = urllib.parse.urlparse(REDIRECT_URI).netloc.split(':')
-
-    if AUTHCODE:
-        return AUTHCODE
-    else:
-        with http.server.HTTPServer((host, int(port)), WebServer) as httpd:
-            httpd = secureServer(httpd, host)
-            while not AUTHCODE:
-                httpd.handle_request()
-    
-    return AUTHCODE
-
-def getTokens():
-    """Exchange auth code for access token"""
-    parser = configparser.ConfigParser()
-    parser.read(configfile)
-
-    tokenEndpoint = 'https://accounts.spotify.com/api/token'
-    grant_type = 'authorization_code'
-    
-    encodedParams = urllib.parse.urlencode({
-        'grant_type': grant_type,
-        'code': AUTHCODE,
-        'redirect_uri': REDIRECT_URI,
-        'client_id': parser.get('spotifyutils', 'client_id'),
-        'client_secret': parser.get('spotifyutils', 'client_secret')
-    }).encode('ascii')
-
-    request = urllib.request.Request(tokenEndpoint, encodedParams)
-    response = urllib.request.urlopen(request).read()
-    json_data = json.loads(response)
-    access_token = json_data['access_token']
-    refresh_token = json_data['refresh_token']
+    print("Getting Auth and Refresh Tokens")
+    getTokens(client_id, client_secret)
