+++ spotifyutils/__main__.py
@@ -1,7 +1,5 @@
-from spotifyutils import cli, server, getTokens
+from spotifyutils import cli
 
 
 if __name__ == '__main__':
     cli()
-    server()
-    getTokens()
